requirements.txt
python-binance
requests
python-dotenv
rich